Editing: API, form refactor, edit toggle
